+++ galite-core/src/main/kotlin/org/kopi/galite/chart/DimensionData.kt
@@ -15,9 +15,9 @@
  * Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301  USA
  */
 
-package org.kopi.galite.visual.chart
+package org.kopi.galite.chart
 
-import org.kopi.galite.visual.field.Field
+import org.kopi.galite.field.Field
 
 /**
  * Represents the value of a dimension
