+++ galite-core/src/main/kotlin/org/kopi/galite/chart/Utils.kt
@@ -15,9 +15,9 @@
  * Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301  USA
  */
 
-package org.kopi.galite.visual.chart
+package org.kopi.galite.chart
 
-import org.kopi.galite.visual.exceptions.MissingMeasureException
+import org.kopi.galite.exceptions.MissingMeasureException
 
 /**
  * Chart data formatter utility.
