+++ galite-tests/src/test/kotlin/org/kopi/galite/tests/chart/MeasureTests.kt
@@ -15,18 +15,23 @@
  * Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301  USA
  */
 
-package org.kopi.galite.visual.chart
+package org.kopi.galite.tests.chart
 
-import org.kopi.galite.domain.Domain
-import org.kopi.galite.visual.common.Color
-import org.kopi.galite.visual.field.Field
+import org.junit.Test
+import org.kopi.galite.chart.Measure
+import org.kopi.galite.visual.Color
+import kotlin.test.assertEquals
 
-/**
- * Represents a measure used to store numeric values in chart.
- *
- * @param domain dimension domain.
- */
-open class Measure<T>(domain: Domain<T>? = null) : Field<T>(domain) where T : Comparable<T>, T : Number {
-  /**Measure's color in chart */
-  lateinit var color: Color
+class MeasureTests {
+
+  /**
+   * Test measure class
+   */
+  @Test
+  fun testMeasure() {
+    val measure1 = Measure<Int>()
+    measure1.label = "measure 1"
+    measure1.color = Color.RED
+    assertEquals(measure1.color.toString(), "RED")
+  }
 }
