+++ galite-tests/src/test/kotlin/org/kopi/galite/tests/domain/CodeDomainTests.kt
@@ -19,8 +19,7 @@ package org.kopi.galite.tests.domain
 
 import org.junit.Test
 import org.kopi.galite.domain.Domain
-import org.kopi.galite.visual.field.Field
-import org.kopi.galite.visual.exceptions.InvalidValueException
+import org.kopi.galite.field.Field
 
 import kotlin.test.assertEquals
 import kotlin.test.assertFailsWith
